pyproject.toml
[project]
name = "hdbscan-app"
version = "0.1.0"
description = "HDBSCAN clustering service for Supabase"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "hdbscan==0.8.40",
    "numpy==2.1.1",
    "supabase==2.6.0",
    "python-dotenv==1.0.1",
]

[tool.uv]
dev-dependencies = []

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
